# openapi.yaml
openapi: 3.0.3
info:
  title: BlockSign API
  description: |
    BlockSign is a secure document signing platform with blockchain anchoring.
    
    ## Authentication
    The API uses Ed25519 challenge-response authentication:
    1. Request a challenge via /auth/challenge
    2. Sign the challenge with your Ed25519 private key
    3. Complete authentication via /auth/complete to receive JWT tokens
    
    Protected endpoints require a Bearer token in the Authorization header.
    
    ## Document Signing Flow
    1. Owner creates document with PDF and list of participants
    2. Participants receive email notification
    3. Each participant signs the canonical payload with their Ed25519 key
    4. When all required signatures are collected, document is anchored to Polygon blockchain
  version: 1.0.0
  contact:
    name: BlockSign Support
    email: blocksigncloud@gmail.com
  license:
    name: Alexei Pavlovschii, BlockSign. All rights reserved.
servers:
  - url: https://api.blocksign.md/api/v1
    description: Production server
  - url: http://localhost:4000/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/complete or /auth/refresh
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid or expired token"
...
paths:
...
  /auth/challenge:
    post:
      tags: [Auth]
      summary: Request login challenge
      description: Generates a random challenge for Ed25519 signature authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
      responses:
        '200':
          description: Challenge generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
                    description: Base64url-encoded random challenge (valid for 5 minutes)
                    example: "dGhpcyBpcyBhIHJhbmRvbSBjaGFsbGVuZ2U"
        '404':
          description: User not found or public key not set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/complete:
    post:
      tags: [Auth]
      summary: Complete authentication
      description: Verifies Ed25519 signature and issues JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, challenge, signatureB64]
              properties:
                email:
                  type: string
                  format: email
                challenge:
                  type: string
                  description: The challenge received from /auth/challenge
                signatureB64:
                  type: string
                  description: Base64-encoded Ed25519 signature of the challenge
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: JWT access token (15 min expiry)
                  refreshToken:
                    type: string
                    description: JWT refresh token (7 day expiry)
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      fullName:
                        type: string
                      role:
                        type: string
        '400':
          description: Invalid or expired challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Uses refresh token from cookie to issue new token pair
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /registration/request/start:
    post:
      tags: [Registration]
      summary: Start registration (send OTP)
      description: Sends a 6-digit OTP to the provided email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /registration/request/verify:
    post:
      tags: [Registration]
      summary: Verify OTP
      description: Verifies the 6-digit OTP sent to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
                  example: "123456"
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /registration/request:
    post:
      tags: [Registration]
      summary: Submit registration request
      description: Submits user details for admin approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, fullName, phone, idnp, username]
              properties:
                email:
                  type: string
                  format: email
                fullName:
                  type: string
                  minLength: 1
                  maxLength: 120
                phone:
                  type: string
                  minLength: 5
                  maxLength: 20
                idnp:
                  type: string
                  minLength: 5
                  maxLength: 20
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                  pattern: '^[a-zA-Z0-9._-]+$'
      responses:
        '201':
          description: Request submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                    format: uuid
        '409':
          description: Request already exists or username taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /registration/complete:
    post:
      tags: [Registration]
      summary: Complete registration
      description: Finalizes registration with Ed25519 public key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, publicKeyEd25519Hex, signatureB64]
              properties:
                token:
                  type: string
                  description: Token from finalization email
                publicKeyEd25519Hex:
                  type: string
                  minLength: 64
                  description: Ed25519 public key in hex format
                signatureB64:
                  type: string
                  description: Base64 signature of the token
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      fullName:
                        type: string
        '400':
          description: Invalid token or signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== USER ====================
...